--- libbeat/reader/readjson/docker_json_common.go
@@ -1,31 +1,0 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-// +build !windows
-
-package readjson
-
-import (
-	"github.com/elastic/beats/libbeat/reader"
-)
-
-func stripNewLine(msg *reader.Message) {
-	l := len(msg.Content)
-	if l > 0 && msg.Content[l-1] == '\n' {
-		msg.Content = msg.Content[:l-1]
-	}
-}
